requirements.txt
asgiref==3.7.2
attrs==23.2.0
blinker==1.6.3
Brotli==1.1.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
Flask==3.0.0
Flask-Script==2.0.6
Flask-WTF==1.2.1
fonttools==4.45.0
gunicorn==22.0.0
hypothesis==6.98.17
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
libsass==0.22.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mock==5.1.0
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.4
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pluggy==1.4.0
protobuf==4.21.12
psycopg2-binary==2.9.9
pycparser==2.21
pygame==2.5.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytest==8.0.2
python-dateutil==2.8.2
QT-PyQt-PySide-Custom-Widgets==0.6.8
QtPy==2.4.1
qtsass==0.4.0
reportlab==4.0.7
requests==2.31.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
WTForms==3.1.1
